Renamed writables folder to stores
